Make setcursor posix complient
Update shebang to /bin/sh

Fix the following in shellcheck
* SC2039: In POSIX sh, [[ ]] is undefined.
* SC2039: In POSIX sh, &> is undefined.
* SC2039: In POSIX sh, -nt is undefined.
* SC2086: Double quote to prevent globbing and word splitting.

Replaced non posix "-nt" and lines that set and check cursor and newest
with a single set and a case.

Additionally added an exit as the case default for case if for some
reason none of the files are found.

Signed-off-by: Malcolm VanOrder <mvanorder1390@gmail.com>